admin product and user manage

+++ Project/Loothub/src/main/java/com/company/springboot/services/IUserService.java
@@ -8,6 +8,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 public interface IUserService extends UserDetailsService{
     public List<User> listAll();
     public  User save(UserRegistrationDto registrationDto);
-    public User get(Long id);
-    public void delete(Long id);
+    public User get(int id);
+    public void delete(int id);
 }
